Address review feedback on coverage tests
